putDec(unsigned short)
git-svn-id: https://micro-os-plus.svn.sourceforge.net/svnroot/micro-os-plus/trunk/Eclipse@1939 ba4b06f3-5e2e-0410-8656-f0fcc57e5dc0

+++ samples/realTimeInterrupts/src/TaskBlink.cpp
@@ -40,7 +40,7 @@ TaskBlink::taskMain(void)
           debug.putString("Task '");
           debug.putString(getName());
           debug.putString("', led=");
-          debug.putDec(m_oLed.bitNumber());
+          debug.putDec((unsigned short)m_oLed.bitNumber());
           debug.putString(", ticks=");
           debug.putDec(m_rate);
           debug.putNewLine();
